Don't ignore errors when running mysql commands in db:start task. That had been done under the assumption that it was most likely failing because it had been already run but there are many reasons for it possibly failing, the most common seems to be mysql not started yet because it took longer than expected to start up.
The commands all look like they can be run more than once without ill effect anyway, but I haven't tested that.

But more importantly, if they fail we should know about it and stop the deploy. And if they're failing from being run twice then we should fix the cause of that (check if db exists and exit, or ensure drop is run before create, for example).